# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Batoi Bolsa  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cyclesPaginate": {
            "get": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Obtener lista de ciclos",
                "description": "Devuelve una lista paginada de ciclos",
                "operationId": "getCyclesIndex",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total_records": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Nombre del ciclo"
                                                    },
                                                    "idFamilia": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "idResponsable": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "links": {
                                            "properties": {
                                                "prev": {
                                                    "type": "string",
                                                    "example": "url_pagina_anterior"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "url_pagina_siguiente"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/cycles/{id}": {
            "get": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Obtener detalles de un ciclo",
                "description": "Devuelve los detalles de un ciclo específico",
                "operationId": "getCycle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ciclo a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Nombre del ciclo"
                                        },
                                        "idFamilia": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "idResponsable": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ciclo no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Actualizar un ciclo existente",
                "description": "Actualiza un ciclo existente con la información proporcionada",
                "operationId": "updateCycle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ciclo a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos actualizados del ciclo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cycle",
                                    "title",
                                    "idFamily",
                                    "idResponsible",
                                    "vliteral",
                                    "cliteral"
                                ],
                                "properties": {
                                    "cycle": {
                                        "type": "string",
                                        "example": "2023-2024"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Título actualizado del ciclo"
                                    },
                                    "idFamily": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "idResponsible": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vliteral": {
                                        "type": "string",
                                        "example": "Valor literal actualizado del ciclo"
                                    },
                                    "cliteral": {
                                        "type": "string",
                                        "example": "Color literal actualizado del ciclo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ciclo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cycle": {
                                            "type": "string",
                                            "example": "2023-2024"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Título actualizado del ciclo"
                                        },
                                        "idFamily": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "idResponsible": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "vliteral": {
                                            "type": "string",
                                            "example": "Valor literal actualizado del ciclo"
                                        },
                                        "cliteral": {
                                            "type": "string",
                                            "example": "Color literal actualizado del ciclo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ciclo no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Eliminar un ciclo existente",
                "description": "Elimina un ciclo existente según su ID",
                "operationId": "deleteCycle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del ciclo a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ciclo eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El ciclo con id: 1 ha sido borrado con éxito"
                                        },
                                        "data": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ciclo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No se ha encontrado el ciclo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/cycles": {
            "get": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Obtener todos los ciclos",
                "description": "Obtiene todos los ciclos disponibles",
                "operationId": "getAllCyclesSinCollction",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cycle"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Crear un nuevo ciclo",
                "description": "Crea un nuevo ciclo con la información proporcionada",
                "operationId": "createCycle",
                "requestBody": {
                    "description": "Datos del nuevo ciclo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cycle",
                                    "title",
                                    "idFamily",
                                    "idResponsible",
                                    "vliteral",
                                    "cliteral"
                                ],
                                "properties": {
                                    "cycle": {
                                        "type": "string",
                                        "example": "2023-2024"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Título del ciclo"
                                    },
                                    "idFamily": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "idResponsible": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vliteral": {
                                        "type": "string",
                                        "example": "Valor literal del ciclo"
                                    },
                                    "cliteral": {
                                        "type": "string",
                                        "example": "Color literal del ciclo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ciclo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cycle": {
                                            "type": "string",
                                            "example": "2023-2024"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Título del ciclo"
                                        },
                                        "idFamily": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "idResponsible": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "vliteral": {
                                            "type": "string",
                                            "example": "Valor literal del ciclo"
                                        },
                                        "cliteral": {
                                            "type": "string",
                                            "example": "Color literal del ciclo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cycle": {
                "title": "Cycle",
                "description": "Esquema para los datos de un ciclo",
                "required": [
                    "cycle",
                    "title",
                    "id_family",
                    "id_responsible",
                    "vliteral",
                    "cliteral"
                ],
                "properties": {
                    "cycle": {
                        "description": "Descripción del ciclo",
                        "type": "string",
                        "example": "cycle_value"
                    },
                    "title": {
                        "description": "Título del ciclo",
                        "type": "string",
                        "example": "title_value"
                    },
                    "id_family": {
                        "description": "ID de la familia del ciclo",
                        "type": "integer",
                        "example": 1
                    },
                    "id_responsible": {
                        "description": "ID del responsable del ciclo",
                        "type": "integer",
                        "example": 1
                    },
                    "vliteral": {
                        "description": "Valor literal del ciclo",
                        "type": "string",
                        "example": "vliteral_value"
                    },
                    "cliteral": {
                        "description": "Código literal del ciclo",
                        "type": "string",
                        "example": "cliteral_value"
                    }
                },
                "type": "object"
            },
...
        }
    }
}